refactor(manifest): consolidate manifest parsing into manifest_parser.py
UniversalManifestSlicer (7-ecosystem manifest parsing + physical package
location) previously lived in sbom_recorder.py, parallel to the narrower
ManifestParser — two independently-written parsers for the same job.
manifest_parser.py is now the single canonical module for dependency
declaration parsing and on-disk package location; sbom_recorder re-imports
the class so all existing import paths, patch targets, and tests work
unchanged. Pure relocation: class body is byte-for-byte identical, no
behavior change.

Groundwork for the dependency-audit hash cache (PR A of the overhaul
replacing the capped physical audit).
