+++ gitgalaxy/recorders/sbom_recorder.py
@@ -6,18 +6,15 @@
 #          across multiple language ecosystems (NPM, Composer, PyPI, Cargo).
 # ==============================================================================
 
-# galaxyscope:ignore sec_high_risk_execution, sec_io
-
 # galaxyscope:ignore sec_high_risk_execution, ai_guardrails, sec_io
 
 import os
 import json
 import uuid
-import re
 import logging
 from datetime import datetime, timezone
 from pathlib import Path
-from typing import Dict, Tuple, List, Any
+from typing import Dict, List, Any
 
 # Import exclusively from the GitGalaxy Hub
 from gitgalaxy.security.security_lens import SecurityLens
@@ -26,158 +23,10 @@ from gitgalaxy.standards.language_standards import LANGUAGE_DEFINITIONS
 from gitgalaxy.standards.gitgalaxy_config import LEXICAL_FAMILY_HEURISTICS
 from gitgalaxy.standards.language_lens import LanguageDetector
 
-
-class UniversalManifestSlicer:
-    """Uses regex and standard parsing to slice dependencies from any ecosystem."""
-
-    @staticmethod
-    def slice_manifest(manifest_path: Path) -> Tuple[str, Dict[str, str]]:
-        filename = manifest_path.name
-        deps = {}
-        ecosystem = "unknown"
-
-        try:
-            if filename == "package.json":
-                ecosystem = "npm"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    data = json.load(f)
-                    deps.update(data.get("dependencies", {}))
-                    deps.update(data.get("devDependencies", {}))
-
-            elif filename == "composer.json":
-                ecosystem = "packagist"  # PHP Composer
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    data = json.load(f)
-                    deps.update(data.get("require", {}))
-                    deps.update(data.get("require-dev", {}))
-                    # Remove the php version requirement
-                    deps.pop("php", None)
-
-            elif filename == "requirements.txt":
-                ecosystem = "pypi"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    for line in f:
-                        line = line.strip()
-                        if line and not line.startswith("#"):
-                            clean_name = re.split(r"[=><~]", line)[0].strip()
-                            # Get version if explicitly defined, else 'latest'
-                            version = line.split("==")[1].strip() if "==" in line else "latest"
-                            deps[clean_name] = version
-
-            elif filename == "Cargo.toml":
-                ecosystem = "cargo"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    content = f.read()
-                    # Universal regex to grab [dependencies] blocks
-                    dep_blocks = re.findall(r"\[(?:dev-)?dependencies\](.*?)(\n\[|$)", content, re.DOTALL)
-                    for block, _ in dep_blocks:
-                        for line in block.splitlines():
-                            line = line.strip()
-                            if line and not line.startswith("#") and "=" in line:
-                                pkg_name = line.split("=")[0].strip()
-                                deps[pkg_name] = "latest"  # Simplified version extraction
-
-            elif filename == "go.mod":
-                ecosystem = "golang"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    in_require_block = False
-                    for line in f:
-                        line = line.strip()
-                        if line.startswith("require ("):
-                            in_require_block = True
-                            continue
-                        if line == ")":
-                            in_require_block = False
-                            continue
-                        if in_require_block and line and not line.startswith("//"):
-                            parts = line.split()
-                            if len(parts) >= 2:
-                                deps[parts[0]] = parts[1]
-                        elif line.startswith("require ") and not in_require_block:
-                            parts = line.split()
-                            if len(parts) >= 3:
-                                deps[parts[1]] = parts[2]
-
-            elif filename == "Gemfile":
-                ecosystem = "rubygems"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    for line in f:
-                        line = line.strip()
-                        # Extract: gem 'nokogiri', '~> 1.11'
-                        if line.startswith("gem "):
-                            parts = line.split(",")
-                            pkg_name = parts[0].replace("gem", "").strip(" '\"")
-                            version = parts[1].strip(" '\"") if len(parts) > 1 else "latest"
-                            deps[pkg_name] = version
-
-            elif filename == "pom.xml":
-                ecosystem = "maven"
-                with open(manifest_path, "r", encoding="utf-8") as f:
-                    content = f.read()
-                    # Extract artifactId and version from XML blocks
-                    deps_raw = re.findall(
-                        r"<dependency>.*?<artifactId>([^<]+)</artifactId>(?:.*?<version>([^<]+)</version>)?.*?</dependency>",
-                        content,
-                        re.DOTALL,
-                    )
-                    for artifact, version in deps_raw:
-                        deps[artifact] = version if version else "latest"
-
-        except Exception:
-            pass
-
-        return ecosystem, deps
-
-    @staticmethod
-    def locate_physical_package(target_path: Path, pkg_name: str, ecosystem: str) -> Path:
-        """Hunts for the physical location of a package within the project bounds."""
-        if ecosystem == "npm":
-            target = target_path / "node_modules" / pkg_name
-            return target if target.exists() else None
-
-        elif ecosystem == "packagist":
-            # Composer packages are in vendor/vendor-name/package-name
-            target = target_path / "vendor" / pkg_name
-            return target if target.exists() else None
-
-        elif ecosystem == "pypi":
-            # Hunt through common local virtual environment folders
-            safe_pkg_name = pkg_name.replace("-", "_").lower()
-            for venv_dir in ["venv", ".venv", "env"]:
-                venv_path = target_path / venv_dir
-                if venv_path.exists():
-                    for root, dirs, _ in os.walk(venv_path):
-                        if "site-packages" in root:
-                            # Case-insensitive match for the package folder
-                            for d in dirs:
-                                if d.lower() == safe_pkg_name or d.lower().startswith(f"{safe_pkg_name}-"):
-                                    return Path(root) / d
-
-        elif ecosystem == "golang":
-            # Go packages are often vendored in the project's root 'vendor/' directory
-            target = target_path / "vendor" / pkg_name
-            return target if target.exists() else None
-
-        elif ecosystem == "rubygems":
-            # Ruby often vendors gems locally here
-            target = target_path / "vendor" / "bundle"
-            if target.exists():
-                for root, dirs, _ in os.walk(target):
-                    if pkg_name in dirs:
-                        return Path(root) / pkg_name
-            return None
-
-        elif ecosystem == "maven":
-            # Java local dependency pulls usually land here
-            target = target_path / "target" / "dependency"
-            if target.exists():
-                # Just verifying the jar/folder exists loosely
-                for file in target.iterdir():
-                    if pkg_name.lower() in file.name.lower():
-                        return file
-            return None
-
-        return None
+# UniversalManifestSlicer now lives in the canonical manifest module (PR A of
+# the dependency-audit overhaul). Re-imported here so existing consumers and
+# tests importing it from this module keep working unchanged.
+from gitgalaxy.security.manifest_parser import UniversalManifestSlicer
 
 
 class SbomRecorder:
@@ -248,13 +97,18 @@ class SbomRecorder:
                     anomaly_notes.append("Package declared in manifest but not found locally.")
                     total_missing += 1
                 else:
-                    # Physical Audit (Max 5 core files to maintain velocity)
-                    scanned_files = 0
+                    # Physical Audit (Max 5 core files PER DIRECTORY to
+                    # maintain velocity). The cap is reset on every directory
+                    # os.walk visits, and enforced per-file (not once per
+                    # directory) so a single bloated folder can't consume the
+                    # whole package's audit budget and starve every other
+                    # directory from being sampled at all (#254).
                     for root, _, files in os.walk(pkg_path):
-                        if scanned_files > 5:
-                            break
-
+                        scanned_in_dir = 0
                         for file in files:
+                            if scanned_in_dir >= 5:
+                                break
+
                             if not file.endswith((".js", ".py", ".ts", ".php", ".rs")):
                                 continue
 
@@ -273,10 +127,9 @@ class SbomRecorder:
                                     trust_status = "SPOOF_DETECTED"
                                     anomaly_notes.extend(id_result["anomaly_flags"])
 
-                                scanned_files += 1
+                                scanned_in_dir += 1
                             except Exception as e:
                                 self.logger.debug(f"Skipped unreadable file during physical audit ({file_path}): {e}")
-
                     if trust_status == "SPOOF_DETECTED":
                         total_anomalies += 1
                         self.logger.warning(f"🚨 [SPOOF DETECTED] {pkg_name}@{pkg_version}")
