+++ gitgalaxy/security/manifest_parser.py
@@ -6,8 +6,10 @@
 # ==============================================================================
 import json
 import logging
+import os
 import re
 from pathlib import Path
+from typing import Dict, Tuple
 
 
 class ManifestParser:
@@ -174,3 +176,165 @@ class ManifestParser:
                             self.logger.warning(f"🚨 Manifest Parser: INSECURE REGISTRY PROTOCOL DETECTED -> {url}")
                             # Prefix with INSECURE_REGISTRY so the Supply Chain Firewall can instantly block it
                             resolution_map[f"INSECURE_REGISTRY_{filepath.name}"] = url
+
+
+class UniversalManifestSlicer:
+    """
+    Uses regex and standard parsing to slice dependencies from any ecosystem.
+
+    RELOCATED (PR A of the dependency-audit overhaul): this class previously
+    lived in gitgalaxy.recorders.sbom_recorder, creating a second, parallel
+    manifest parser alongside ManifestParser above. manifest_parser.py is now
+    the single canonical module for "what does this project depend on, and
+    where does it physically live on disk." sbom_recorder re-imports this
+    class for backward compatibility.
+    """
+
+    @staticmethod
+    def slice_manifest(manifest_path: Path) -> Tuple[str, Dict[str, str]]:
+        filename = manifest_path.name
+        deps = {}
+        ecosystem = "unknown"
+
+        try:
+            if filename == "package.json":
+                ecosystem = "npm"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    data = json.load(f)
+                    deps.update(data.get("dependencies", {}))
+                    deps.update(data.get("devDependencies", {}))
+
+            elif filename == "composer.json":
+                ecosystem = "packagist"  # PHP Composer
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    data = json.load(f)
+                    deps.update(data.get("require", {}))
+                    deps.update(data.get("require-dev", {}))
+                    # Remove the php version requirement
+                    deps.pop("php", None)
+
+            elif filename == "requirements.txt":
+                ecosystem = "pypi"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    for line in f:
+                        line = line.strip()
+                        if line and not line.startswith("#"):
+                            clean_name = re.split(r"[=><~]", line)[0].strip()
+                            # Get version if explicitly defined, else 'latest'
+                            version = line.split("==")[1].strip() if "==" in line else "latest"
+                            deps[clean_name] = version
+
+            elif filename == "Cargo.toml":
+                ecosystem = "cargo"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    content = f.read()
+                    # Universal regex to grab [dependencies] blocks
+                    dep_blocks = re.findall(r"\[(?:dev-)?dependencies\](.*?)(\n\[|$)", content, re.DOTALL)
+                    for block, _ in dep_blocks:
+                        for line in block.splitlines():
+                            line = line.strip()
+                            if line and not line.startswith("#") and "=" in line:
+                                pkg_name = line.split("=")[0].strip()
+                                deps[pkg_name] = "latest"  # Simplified version extraction
+
+            elif filename == "go.mod":
+                ecosystem = "golang"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    in_require_block = False
+                    for line in f:
+                        line = line.strip()
+                        if line.startswith("require ("):
+                            in_require_block = True
+                            continue
+                        if line == ")":
+                            in_require_block = False
+                            continue
+                        if in_require_block and line and not line.startswith("//"):
+                            parts = line.split()
+                            if len(parts) >= 2:
+                                deps[parts[0]] = parts[1]
+                        elif line.startswith("require ") and not in_require_block:
+                            parts = line.split()
+                            if len(parts) >= 3:
+                                deps[parts[1]] = parts[2]
+
+            elif filename == "Gemfile":
+                ecosystem = "rubygems"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    for line in f:
+                        line = line.strip()
+                        # Extract: gem 'nokogiri', '~> 1.11'
+                        if line.startswith("gem "):
+                            parts = line.split(",")
+                            pkg_name = parts[0].replace("gem", "").strip(" '\"")
+                            version = parts[1].strip(" '\"") if len(parts) > 1 else "latest"
+                            deps[pkg_name] = version
+
+            elif filename == "pom.xml":
+                ecosystem = "maven"
+                with open(manifest_path, "r", encoding="utf-8") as f:
+                    content = f.read()
+                    # Extract artifactId and version from XML blocks
+                    deps_raw = re.findall(
+                        r"<dependency>.*?<artifactId>([^<]+)</artifactId>(?:.*?<version>([^<]+)</version>)?.*?</dependency>",
+                        content,
+                        re.DOTALL,
+                    )
+                    for artifact, version in deps_raw:
+                        deps[artifact] = version if version else "latest"
+
+        except Exception:
+            pass
+
+        return ecosystem, deps
+
+    @staticmethod
+    def locate_physical_package(target_path: Path, pkg_name: str, ecosystem: str) -> Path:
+        """Hunts for the physical location of a package within the project bounds."""
+        if ecosystem == "npm":
+            target = target_path / "node_modules" / pkg_name
+            return target if target.exists() else None
+
+        elif ecosystem == "packagist":
+            # Composer packages are in vendor/vendor-name/package-name
+            target = target_path / "vendor" / pkg_name
+            return target if target.exists() else None
+
+        elif ecosystem == "pypi":
+            # Hunt through common local virtual environment folders
+            safe_pkg_name = pkg_name.replace("-", "_").lower()
+            for venv_dir in ["venv", ".venv", "env"]:
+                venv_path = target_path / venv_dir
+                if venv_path.exists():
+                    for root, dirs, _ in os.walk(venv_path):
+                        if "site-packages" in root:
+                            # Case-insensitive match for the package folder
+                            for d in dirs:
+                                if d.lower() == safe_pkg_name or d.lower().startswith(f"{safe_pkg_name}-"):
+                                    return Path(root) / d
+
+        elif ecosystem == "golang":
+            # Go packages are often vendored in the project's root 'vendor/' directory
+            target = target_path / "vendor" / pkg_name
+            return target if target.exists() else None
+
+        elif ecosystem == "rubygems":
+            # Ruby often vendors gems locally here
+            target = target_path / "vendor" / "bundle"
+            if target.exists():
+                for root, dirs, _ in os.walk(target):
+                    if pkg_name in dirs:
+                        return Path(root) / pkg_name
+            return None
+
+        elif ecosystem == "maven":
+            # Java local dependency pulls usually land here
+            target = target_path / "target" / "dependency"
+            if target.exists():
+                # Just verifying the jar/folder exists loosely
+                for file in target.iterdir():
+                    if pkg_name.lower() in file.name.lower():
+                        return file
+            return None
+
+        return None
